ISSUE #3021 - moved groups to tree to a helper
